Move over to new chats page and remove docs pages
Update tests

Update tests

Apply fixes from tests

Update model-selector test

Update test
