+++ django_app/frontend/src/css/chats/chats.scss
@@ -1,5 +1,5 @@
-main:has(.iai-chat-input),
-body:has(.iai-environment-warning) main:has(.iai-chat-input) {
+main:has(.rb-chat-input),
+body:has(.iai-environment-warning) main:has(.rb-chat-input) {
   min-height: 0;
 }
 
@@ -17,7 +17,7 @@ body:has(.iai-environment-warning) main:has(.iai-chat-input) {
   .rb-chats-section {
     flex-grow: 1;
   }
-  .govuk-grid-row:has(.iai-chat-input) {
+  .govuk-grid-row:has(.rb-chat-input) {
     display: flex;
     .govuk-grid-column-one-third,
     .govuk-grid-column-two-thirds {
@@ -35,7 +35,7 @@ body:has(.iai-environment-warning) main:has(.iai-chat-input) {
   }
 }
 @media (min-width: 1020px) {
-  .govuk-grid-row:has(.iai-chat-input) {
+  .govuk-grid-row:has(.rb-chat-input) {
     .govuk-grid-column-one-third,
     .govuk-grid-column-two-thirds {
       min-height: calc(100vh - 10.25rem);
@@ -43,12 +43,12 @@ body:has(.iai-environment-warning) main:has(.iai-chat-input) {
   }
 }
 @media (min-width: 1020px) and (min-height: 600px) {
-  body:has(.iai-chat-input) .iai-environment-warning {
+  body:has(.rb-chat-input) .iai-environment-warning {
     position: fixed;
     width: 100%;
     z-index: 10;
   }
-  body:has(.iai-chat-input) .iai-top-nav {
+  body:has(.rb-chat-input) .iai-top-nav {
     position: fixed;
     width: 100%;
     z-index: 10;
@@ -56,10 +56,10 @@ body:has(.iai-environment-warning) main:has(.iai-chat-input) {
   body:has(.iai-environment-warning) .iai-top-nav {
     top: 26px;
   }
-  main:has(.iai-chat-input) {
+  main:has(.rb-chat-input) {
     padding-top: 4.5rem;
   }
-  body:has(.iai-environment-warning) main:has(.iai-chat-input) {
+  body:has(.iai-environment-warning) main:has(.rb-chat-input) {
     padding-top: 6.5rem;
   }
   .rb-left-column-container {
@@ -70,12 +70,6 @@ body:has(.iai-environment-warning) main:has(.iai-chat-input) {
   }
   body:has(.iai-environment-warning) .rb-left-column-container {
     top: 6.5rem;
-  }
-  .iai-panel__scrollable {
-    max-height: calc(50vh - 12rem);
-  }
-  main:has(#llm-selector) .iai-panel__scrollable {
-    max-height: calc(50vh - 15rem);
   }
 }
 
@@ -142,7 +136,7 @@ chat-message {
 .rb-chat-message__container {
   margin-top: -1rem;
 }
-.govuk-main-wrapper:has(.iai-chat-input) {
+.govuk-main-wrapper:has(.rb-chat-input) {
   padding-bottom: 0;
 }
 
@@ -212,13 +206,6 @@ chat-message {
   background-color: $redbox-red;
 }
 
-.iai-chat-input {
-  max-width: 94%;
-}
-@media (min-width: 1020px) {
-  .iai-chat-input {
-    max-width: 100%;
-  }
-}
+@import "./chat-input.scss";
 
 @import "./documents.scss";
