+++ django_app/frontend/src/css/chats/model-selector.scss
@@ -1,21 +1,30 @@
 
 .rb-model-selector {
   cursor: pointer;
+  display: inline-block;
+  margin-left: 1rem;
   position: relative;
 }
 .rb-model-selector__select {
   align-items: center;
   display: flex;
+  font-size: 0.75rem;
   justify-content: space-between;
+}
+.rb-model-selector__select span {
+  display: inline-block;
+  font-size: 0.5rem;
+  margin-left: 0.5rem;
 }
 .rb-model-selector__list {
   background-color: white;
-  border: 2px solid;
-  bottom: 2.5rem;
+  border: 1px solid;
+  bottom: 1rem;
+  font-size: 0.75rem;
   left: -0.5rem;
+  min-width: 14rem;
   padding: 0.5rem;
   position: absolute;
-  width: 150%;
 }
 .rb-model-selector__option {
   padding: 0.5rem;
@@ -26,6 +35,8 @@
 .rb-model-selector__option span {
   display: block;
   pointer-events: none;
+  word-wrap: none;
+  word-break: keep-all;
 }
 .rb-model-selector__option span:first-child {
   font-weight: 700;
