+++ django_app/frontend/src/css/print.scss
@@ -14,7 +14,7 @@
   }
 
   /* Hide other interactive elements */
-  .chat-title__edit-btn, copy-text, feedback-buttons, .exit-feedback, .iai-chat-input {
+  .chat-title__edit-btn, copy-text, feedback-buttons, .exit-feedback, .rb-chat-input {
     display: none !important;
   }
