+++ django_app/frontend/src/js/chats.js
@@ -5,18 +5,18 @@ import "./web-components/chats/chat-history-item.js";
 import "./web-components/chats/chat-message.js";
 import "./web-components/chats/chat-title.js";
 import "./web-components/chats/copy-text.js";
-import "./web-components/chats/document-selector.js";
-import "./web-components/chats/document-container.mjs";
-import "./web-components/chats/document-upload.mjs";
 import "./web-components/chats/feedback-buttons.js";
 import "./web-components/chats/markdown-converter.js";
 import "./web-components/chats/message-input.js";
 import "./web-components/chats/model-selector.mjs";
 import "./web-components/chats/canned-prompts";
 import "./web-components/chats/send-message.js";
-import "./web-components/documents/file-status.js";
-import "./web-components/chats/upload-container.mjs";
 import "./web-components/chats/exit-feedback.js";
+
+import "./web-components/documents/document-container.mjs";
+import "./web-components/documents/document-upload.mjs";
+import "./web-components/documents/file-status.js";
+import "./web-components/documents/upload-container.mjs";
 
 document.addEventListener("chat-response-end", (evt) => {
