--- django_app/frontend/src/js/document-upload.js
@@ -1,1 +1,0 @@
-import "./web-components/documents/upload-button.js";
