--- django_app/frontend/src/js/documents.js
@@ -1,2 +1,0 @@
-import "./web-components/documents/doc-list.js";
-import "./web-components/documents/file-status.js";
