+++ django_app/frontend/src/js/web-components/chats/canned-prompts.js
@@ -38,7 +38,7 @@ class CannedPrompts extends HTMLElement {
    */
   #prepopulateMessageBox = (prompt) => {
     /** @type HTMLInputElement | null */
-    let chatInput = document.querySelector(".iai-chat-input__input");
+    let chatInput = document.querySelector(".rb-chat-input textarea");
     if (chatInput) {
       chatInput.value = prompt;
       chatInput.focus();
