+++ django_app/frontend/src/js/web-components/documents/document-container.mjs
