+++ django_app/frontend/src/js/web-components/documents/upload-container.mjs
