+++ django_app/frontend/tests-web-components/canned-prompts.spec.js
@@ -6,7 +6,7 @@ test(`Clicking canned prompts updates the text input`, async ({ page }) => {
 
   await page.goto("/chats");
 
-  const textInput = page.locator(".iai-chat-input__input");
+  const textInput = page.locator(".rb-chat-input textarea");
   await expect(textInput).toHaveValue("");
 
   await page.locator(".chat-options__option").nth(0).click();
