+++ django_app/frontend/tests-web-components/copy-text.spec.js
@@ -11,7 +11,7 @@ test(`Content can be copied to clipboard`, async ({ page }) => {
   await page.getByText("Copy to clipboard").click();
   const response = page.locator(".iai-chat-bubble__text").nth(1);
 
-  const messageInput = page.locator(".iai-chat-input__input");
+  const messageInput = page.locator(".rb-chat-input textarea");
   await messageInput.focus();
   await page.keyboard.press("Meta+V");
   expect(await response.textContent()).toEqual(await messageInput.inputValue());
