--- django_app/redbox_app/templates/documents.html
@@ -1,104 +1,0 @@
-{% set pageTitle = "Documents" %}
-{% extends "base.html" %}
-{% from "macros/govuk-button.html" import govukButton %}
-{% from "macros/iai-doc-list.html" import iaiDocList %}
-
-{% block content %}
-
-<script src="{{ static('js/documents.js') }}"></script>
-
-<div class="govuk-width-container">
-
-  <div class="govuk-grid-row">
-    <div class="govuk-grid-column-two-thirds-from-desktop">
-
-      {% if ingest_errors %}
-        <div class="govuk-error-summary" data-module="govuk-error-summary">
-          <div role="alert">
-            <h2 class="govuk-error-summary__title">
-              There was a problem uploading some documents
-            </h2>
-            <div class="govuk-error-summary__body">
-              <ul class="govuk-list govuk-error-summary__list">
-                {% for error in ingest_errors %}
-                  <li>{{ error }}</li>
-                {% endfor %}
-              </ul>
-            </div>
-          </div>
-        </div>
-      {% endif %}
-
-      <h1 class="govuk-heading-l govuk-!-margin-top-5">Your documents</h1>
-
-      <p class="govuk-body-l">Manage documents to use with your Redbox.</p>
-
-      <p class="govuk-body">Redbox is a service that lets you securely use a Large Language Model (LLM) with or without documents up to OFFICIAL SENSITIVE.</p>
-      <p class="govuk-body govuk-!-margin-bottom-1">This means you can:</p>
-      <ul class="govuk-list govuk-list--bullet govuk-body">
-        <li>Chat directly with an LLM (currently GPT-4o)</li>
-        <li>Upload OFFICIAL and OFFICIAL SENSITIVE documents that do not include personal data.</li>
-        <li>Summarise or ask questions of these documents.</li>
-      </ul>
-
-      {#
-      <div class="rb-doc-retention govuk-!-margin-bottom-5">
-        <img width="32" height="32" src="{{ static('images/icon-30-days.png') }}" loading="lazy" alt=""/>
-        <p class="govuk-body-s govuk-!-margin-bottom-0">Documents will be stored for up to 30 days.</p>
-      </div>
-      #}
-
-      {# New Docs list #}
-      <h2 class="govuk-heading-m govuk-!-margin-top-7">Ready to use</h2>
-      <div class="rb-docs">
-        <p class="rb-docs__show-if-no-docs">Documents that are ready to chat with will be displayed here.</p>
-        <p class="rb-docs__show-if-docs">These documents are ready to chat with. Documents are deleted after 30 days.</p>
-        <doc-list>
-          {{ iaiDocList(
-            docs = completed_files,
-            type = "complete"
-          ) }}
-        </doc-list>
-      </div>
-
-      <h2 class="govuk-heading-m govuk-!-margin-top-7">Processing</h2>
-      <div class="rb-docs">
-        <p class="rb-docs__show-if-no-docs">Uploaded files that are processing will be displayed here.</p>
-        <p class="rb-docs__show-if-docs">These documents will not be included in chat responses.</p>
-        {{ iaiDocList(
-          docs = processing_files,
-          type = "processing"
-        ) }}
-        <div class="rb-docs__upload-error">
-          <svg width="32" height="32" viewBox="0 0 18 18" fill="none" focusable="false" aria-hidden="true">
-            <path d="M7.875 10.125H10.125V4.5H7.875V10.125ZM10.125 12.375C10.125 11.7529 9.62212 11.25 9 11.25C8.37788 11.25 7.875 11.7529 7.875 12.375C7.875 12.9971 8.37788 13.5 9 13.5C9.62212 13.5 10.125 12.9971 10.125 12.375ZM18 9C18 4.02863 13.9714 0 9 0C4.02863 0 0 4.02863 0 9C0 13.9714 4.02863 18 9 18C13.9714 18 18 13.9714 18 9ZM15.75 9C15.75 10.8034 15.0469 12.4976 13.7745 13.7745C12.4976 15.0469 10.8034 15.75 9 15.75C7.19662 15.75 5.50238 15.0469 4.2255 13.7745C2.95312 12.4976 2.25 10.8034 2.25 9C2.25 7.19662 2.95312 5.50238 4.2255 4.2255C5.50238 2.95312 7.19662 2.25 9 2.25C10.8034 2.25 12.4976 2.95312 13.7745 4.2255C15.0469 5.50238 15.75 7.19662 15.75 9Z" fill="currentColor"/>
-          </svg>
-          <p class="govuk-!-margin-0">You have documents that failed to process. Please try again, and if the problem persists contact <a class="govuk-link" href="mailto:">redbox-support@cabinetoffice.gov.uk</a> with details.</p>
-        </div>
-      </div>
-
-      <h2 class="govuk-visually-hidden">General actions</h2>
-      <div class="govuk-button-group govuk-!-margin-top-8">
-        {% if completed_files or processing_files %}
-          {{ govukButton(
-            text="Start chat",
-            href=url('chats')
-          ) }}
-          {{ govukButton(
-            text="Add documents",
-            href=url('upload'),
-            classes="govuk-button--secondary"
-          ) }}
-        {% else %}
-          {{ govukButton(
-            text="Add documents",
-            href=url('upload')
-          ) }}
-        {% endif %}
-      </div>
-
-    </div>
-  </div>
-</div>
-
-{% endblock %}
