--- django_app/redbox_app/templates/remove-doc.html
@@ -1,40 +1,0 @@
-{% set pageTitle = "Remove document" %}
-
-{% extends "base.html" %}
-{% from "macros/govuk-button.html" import govukButton %}
-{% from "macros/govuk-error-summary.html" import govukErrorSummary %}
-
-{% block content %}
-
-<form method="post" enctype="multipart/form-data">
-
-<div class="govuk-width-container">
-
-  <a href="{{url('documents')}}" class="govuk-back-link govuk-!-margin-bottom-5 govuk-!-margin-top-0">Back</a>
-
-  <form method="post" enctype="multipart/form-data">
-
-    <input type="hidden" name="csrfmiddlewaretoken" value="{{ csrf_token }}">
-    <input type="hidden" name="doc_id" value="{{ doc_id }}">
-
-    <h1 class="govuk-heading-m">Are you sure you want to remove this document?</h1>
-
-    {{ govukErrorSummary(error_list=errors) }}
-
-    <dl class="govuk-summary-list">
-      <div class="govuk-summary-list__row">
-        <dt class="govuk-summary-list__key">
-          Document name:
-        </dt>
-        <dd class="govuk-summary-list__value">
-          {{ doc_name }}
-        </dd>
-      </div>
-    </dl>
-
-    {{ govukButton(text="Remove", classes="govuk-button--warning", prevent_double_click=True) }}
-
-  </form>
-</div>
-
-{% endblock %}
