+++ django_app/redbox_app/templates/sitemap.html
@@ -7,31 +7,21 @@
     <div class="govuk-grid-row">
         <div class="govuk-grid-column-two-thirds">
 
-            <h1 class="govuk-heading-l">Sitemap</h1>
+            <h1 class="govuk-heading-l govuk-!-margin-bottom-0 govuk-!-margin-top-3">Sitemap</h1>
 
-            <h2 class="govuk-heading-m govuk-!-margin-top-5 govuk-!-padding-top-5">General pages</h2>
+            <h2 class="govuk-heading-m govuk-!-margin-top-3 govuk-!-padding-top-5">General pages</h2>
             <ul class="govuk-list">
                 <li><a class="govuk-link" href="{{ url('accessibility-statement') }}">Accessibility</a></li>
                 <li><a class="govuk-link" href="{{ url('homepage') }}">Home page</a></li>
+                {% if not request.user.is_authenticated %}
+                    <li><a class="govuk-link" href="{{ url('log-in') }}">Log in</a></li>
+                {% endif %}
                 {% if request.user.is_authenticated %}
                     <li><a class="govuk-link" href="{{ url('demographics') }}">My details</a></li>
                 {% endif %}
                 <li><a class="govuk-link" href="{{ url('privacy-notice') }}">Privacy</a></li>
-                {% if not request.user.is_authenticated %}
-                    <li><a class="govuk-link" href="{{ url('log-in') }}">Log in</a></li>
-                {% endif %}
                 <li><a class="govuk-link" href="{{ url('support') }}">Support</a></li>
             </ul>
-
-            <h2 class="govuk-heading-m govuk-!-margin-top-5 govuk-!-padding-top-5">Documents</h2>
-            {% if request.user.is_authenticated %}
-                <ul class="govuk-list">
-                    <li><a class="govuk-link" href="{{ url('upload') }}">Add documents</a></li>
-                    <li><a class="govuk-link" href="{{ url('documents') }}">Documents</a></li>
-                </ul>
-            {% else %}
-                <p class="govuk-body">These pages are available once you have signed in.</p>
-            {% endif %}
 
             <h2 class="govuk-heading-m govuk-!-margin-top-5 govuk-!-padding-top-5">Chats</h2>
             {% if request.user.is_authenticated %}
