--- django_app/redbox_app/templates/upload.html
@@ -1,48 +1,0 @@
-{% set pageTitle = "Upload a document" %}
-
-{% extends "base.html" %}
-{% from "macros/govuk-button.html" import govukButton %}
-
-{% block content %}
-
-<div class="govuk-width-container">
-  <div class="govuk-grid-row">
-    <div class="govuk-grid-column-two-thirds">
-
-      <a href="{{url('documents')}}" class="govuk-back-link govuk-!-margin-bottom-5 govuk-!-margin-top-0">Back</a>
-
-      <form method="post" enctype="multipart/form-data">
-
-        <input type="hidden" name="csrfmiddlewaretoken" value="{{ csrf_token }}">
-
-        <div class="govuk-form-group {% if errors.upload_doc %} govuk-form-group--error{% endif %}">
-          <label class="govuk-label" for="upload-docs">
-            <h1 class="govuk-heading-l">{{ pageTitle }}</h1>
-          </label>
-          
-          <div id="upload-docs-notification">
-            <p class="govuk-body-l">The AI will use all documents you upload. You can use up to, and including, {{ security }} documents. Do not upload any documents with personal data.</p>
-          </div>
-
-          <p class="govuk-body rb-file-types" id="upload-docs-filetypes">Limit 200MB per file: EML, HTML, JSON, MD, MSG, RST, RTF, TXT, XML, CSV, DOC, DOCX, EPUB, ODT, PDF, PPT, PPTX, TSV, XLSX, HTM</p>
-          {% for error in errors.upload_doc %}
-            <p id="file-upload-docs-error" class="govuk-error-message">
-              <span class="govuk-visually-hidden">Error:</span> {{ error }}
-            </p>
-          {% endfor %}
-          <input class="govuk-file-upload {% if errors.upload_doc %} govuk-file-upload--error{% endif %}" multiple id="upload-docs" name="uploadDocs" type="file" aria-describedby="upload-docs-notification upload-docs-filetypes {% if errors.upload_doc %} file-upload-docs-error{% endif %}">
-        </div>
-
-        <upload-button>
-          {{ govukButton(text="Upload", prevent_double_click=True, classes="govuk-!-display-inline-block") }}
-        </upload-button>
-
-      </form>
-
-    </div>
-  </div>
-</div>
-
-<script src="{{ static('js/document-upload.js') }}" type="module"></script>
-
-{% endblock %}
