+++ django_app/redbox_app/urls.py
@@ -42,7 +42,6 @@ info_urlpatterns = [
 ]
 
 file_urlpatterns = [
-    path("documents/", views.DocumentView.as_view(), name="documents"),
     path("upload/", views.UploadView.as_view(), name="upload"),
     path("upload/<uuid:chat_id>/", views.UploadView.as_view(), name="upload"),
     path("remove-doc/<uuid:doc_id>", views.remove_doc_view, name="remove-doc"),
